Fix umask
As we switched to a SAP compliant image, it turned out that the umask is
set to the restrictive value of `077`. What this does is that when the
root user installs binaries (by creating them under /usr/locla/bin)
those binaries end up as `rwx--x--x`. This then causes the following
error with some binaries

```
Cannot open /proc/self/exe: Permission denied
Failed to get fs offset for /proc/self/exe
```

as the binariy run by a non-root user cannot read itself via
`proc/self/exe`

This is fixed by setting the umask back to the default `022` value

As the /etc/login.defs file takes effect after logging out and then back
in again we have to temporarily set umask to 022 in the provision.sh
script to make sure the installed binaries have `rwxr-xr-x` (755) mode.